refactor: consolidate caching code under src/cache/
Move ArtifactCache (was keys/source_build/cache.rs), WorkspaceCache
(was keys/source_build/workspace.rs), CacheDirs (was cache_dirs.rs),
and rename build_backend_metadata.rs -> backend_metadata.rs so all
on-disk caches live in one module. Adds a module-level rustdoc
overview to cache/mod.rs and cache/{backend_metadata,dirs}.rs.

No public-API changes: the four crate-level re-exports (CacheDirs,
BuildBackendMetadataCache, BuildBackendMetadataCacheEntry,
CacheRevision, MetadataCache, CacheEntry) now go through cache::*.
