+++ crates/pixi_command_dispatcher/src/build_backend_metadata/mod.rs
@@ -17,7 +17,7 @@ use std::{
 use thiserror::Error;
 
 use crate::build::CanonicalSourceCodeLocation;
-use crate::cache::common::CacheRevision;
+use crate::cache::CacheRevision;
 use crate::compute_data::{HasBuildBackendMetadataCache, HasCacheDirs, HasReporter};
 use crate::injected_config::{BackendOverrideKey, EnabledProtocolsKey};
 use crate::input_hash::{ConfigurationHash, ProjectModelHash};
@@ -28,11 +28,9 @@ use crate::{
     InstantiateBackendKey, SourceCheckout, SourceCheckoutError,
     build::{PinnedSourceCodeLocation, SourceRecordOrCheckout, WorkDirKey},
     cache::{
-        build_backend_metadata::{
-            self, BuildBackendMetadataCache, BuildBackendMetadataCacheEntry,
-            BuildBackendMetadataCacheKey,
-        },
-        common::{CacheEntry, CacheKey, CacheKeyString, MetadataCache, MetadataCacheKey},
+        self, BuildBackendMetadataCache, BuildBackendMetadataCacheEntry,
+        BuildBackendMetadataCacheKey, CacheEntry, CacheKey, CacheKeyString, MetadataCache,
+        MetadataCacheKey,
     },
     source_checkout::SourceCheckoutExt,
 };
@@ -875,10 +873,10 @@ impl BuildBackendMetadataInner {
             .await
             .map_err(BuildBackendMetadataError::Cache)?
         {
-            build_backend_metadata::WriteResult::Written => {
+            cache::backend_metadata::WriteResult::Written => {
                 tracing::trace!("Cache updated successfully");
             }
-            build_backend_metadata::WriteResult::Conflict(_other_metadata) => {
+            cache::backend_metadata::WriteResult::Conflict(_other_metadata) => {
                 tracing::debug!(
                     "Cache was updated by another process during computation (version conflict), using our computed result"
                 );
@@ -944,7 +942,7 @@ pub enum BuildBackendMetadataError {
     GlobSet(Arc<pixi_glob::GlobSetError>),
 
     #[error(transparent)]
-    Cache(#[from] build_backend_metadata::BuildBackendMetadataCacheError),
+    Cache(#[from] cache::backend_metadata::BuildBackendMetadataCacheError),
 
     #[error("failed to normalize path")]
     NormalizePath(Arc<pixi_path::NormalizeError>),
