+++ crates/pixi_command_dispatcher/src/cache/artifact.rs
