+++ crates/pixi_command_dispatcher/src/keys/source_build.rs
@@ -5,11 +5,7 @@
 //! gives the backend a stable incremental-build location across runs
 //! sharing the same deps.
 
-mod cache;
-mod workspace;
-
-pub use cache::ArtifactCache;
-pub use workspace::WorkspaceCache;
+pub use crate::cache::{ArtifactCache, WorkspaceCache};
 
 use std::{collections::BTreeMap, hash::Hash, path::PathBuf, sync::Arc};
 
@@ -49,8 +45,7 @@ fn unwrap_dispatcher_err<E>(err: CommandDispatcherError<E>) -> E {
     }
 }
 
-use self::cache::compute_artifact_cache_key;
-use self::workspace::compute_workspace_key;
+use crate::cache::{compute_artifact_cache_key, compute_workspace_key};
 
 /// Hashable inputs to a source build. Runtime concerns (reporters, log
 /// sinks, force-rebuild) stay out of the spec; force-rebuild wipes the
@@ -640,9 +635,9 @@ async fn compute_package_sha256(path: &std::path::Path) -> Result<Sha256Hash, So
     .map_err(|e| SourceBuildError::CalculateSha256(p, Arc::new(e)))
 }
 
-fn map_cache_err(err: self::cache::ArtifactCacheError) -> SourceBuildError {
+fn map_cache_err(err: crate::cache::ArtifactCacheError) -> SourceBuildError {
     match err {
-        self::cache::ArtifactCacheError::Io {
+        crate::cache::ArtifactCacheError::Io {
             operation,
             path,
             source,
@@ -650,8 +645,8 @@ fn map_cache_err(err: self::cache::ArtifactCacheError) -> SourceBuildError {
             let msg = format!("{operation} at {}", path.display());
             SourceBuildError::CreateWorkDirectory(Arc::new(std::io::Error::new(source.kind(), msg)))
         }
-        self::cache::ArtifactCacheError::Glob(err) => SourceBuildError::GlobSet(err),
-        self::cache::ArtifactCacheError::ArtifactFilename(path) => {
+        crate::cache::ArtifactCacheError::Glob(err) => SourceBuildError::GlobSet(err),
+        crate::cache::ArtifactCacheError::ArtifactFilename(path) => {
             SourceBuildError::MissingOutputFile(path)
         }
     }
